Add files via upload
This is a case study project which is part of my Springboard Data Science Career Track curriculum. We ultimately look at which boroughs have largest increase in housing prices over a two decade span. The project was done in Python and makes use of various tools ranging from loops and functions to proper data structures and methods.